+++ src/htpolynet/analysis/postsim.py
@@ -436,7 +436,7 @@ def postsim(args):
     software.sw_setup()
     logger.debug(f'ogromacs {ogromacs}')
     for d in args.proj:
-        pfs.pfs_setup(root=os.getcwd(),topdirs=pfs.Dirs.postsim_topdirs,verbose=True,projdir=d,reProject=False,userlibrary=args.lib)
+        pfs.pfs_setup(root=os.getcwd(),topdirs=pfs.Dirs.postsim_topdirs,verbose=True,projdir=d,reProject=False,userlibrary=pfs.resolve_user_library(args.lib))
         pfs.go_to(pfs.Dirs.postsim)
         for stage in cfg.stagelist:
             stage.do(mdp_pfx='local',**ogromacs)